+++ packages/crypto/docs/classes/PowHelper.md
