+++ packages/crypto/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@iota/crypto.js",
     "description": "IOTA JavaScript Cryptography Library",
-    "version": "1.9.0-stardust.6",
+    "version": "1.9.0-stardust.7",
     "keywords": [
         "iota",
         "client"
