+++ packages/iota/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@iota/iota.js",
     "description": "IOTA JavaScript Library",
-    "version": "1.9.0-stardust.25",
+    "version": "1.9.0-stardust.26",
     "keywords": [
         "iota",
         "client"
@@ -16,7 +16,7 @@
     },
     "license": "Apache-2.0",
     "dependencies": {
-        "@iota/crypto.js": "^1.9.0-stardust.6",
+        "@iota/crypto.js": "^1.9.0-stardust.7",
         "@iota/util.js": "^1.9.0-stardust.5",
         "big-integer": "^1.6.51",
         "node-fetch": "2.6.7"
