+++ packages/pow-node/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@iota/pow-node.js",
     "description": "IOTA JavaScript Library for Performing Multi-Threaded PoW on NodeJS",
-    "version": "1.9.0-stardust.4",
+    "version": "1.9.0-stardust.5",
     "keywords": [
         "iota",
         "pow"
@@ -65,7 +65,7 @@
         "typings"
     ],
     "dependencies": {
-        "@iota/crypto.js": "^1.9.0-stardust.4",
+        "@iota/crypto.js": "^1.9.0-stardust.7",
         "@iota/iota.js": "^1.9.0-stardust.4"
     }
 }
