+++ packages/pow-wasm/README.md
@@ -5,7 +5,7 @@ This package provides local proof of work multi-threaded for NodeJS implemented 
 ## Install
 
 ```shell
-npm install @iota/pow-wasm.js@1.9.0-stardust.4
+npm install @iota/pow-wasm.js@next
 ```
 
 ## Usage
